delete method

+++ src/main/java/com/github/ricardocomar/springbootcamunda/mockservice/entrypoint/model/ScenarioRequest.java
@@ -14,7 +14,7 @@ public class ScenarioRequest {
 
     private ConditionRequest condition;    
 
-    private Long priority;
+    private Integer priority;
     
     private List<VariableRequest> variables;
